+++ common/src/main/java/com/gitlab/srcmc/rctmod/api/config/IModConfig.java
@@ -20,6 +20,8 @@ package com.gitlab.srcmc.rctmod.api.config;
 import net.neoforged.neoforge.common.ModConfigSpec;
 
 public interface IModConfig {
+    static final String SEPARATOR = "━━━━━━━━━━";
+
     /**
      * Reload is invoked after the config was reloaded and all config values have been parsed.
      */